refactor: extract map-contacts to own folder
