+++ src/pages/index/index.pug
@@ -9,18 +9,4 @@ block append main
     include components/about/about
     include components/entertainment-block/entertainment-block
     include components/restaurant-block/restaurant-block
-    .container
-        .map
-            .map-container(data-coords='55.791440, 49.112172')
-            .map-contacts
-                h2.map-contacts__header Как нас найти
-                .map-contacts__container
-                    .map-contacts__content
-                        .map-contacts__content-title Адрес:
-                        .map-contacts__content-text Казань, ул. Баумана, д. 44/8
-                    .map-contacts__content
-                        .map-contacts__content-title Часы работы:
-                        .map-contacts__content-text Ежедневно 8:00-22:00
-                    .map-contacts__button: +anchor('https://yandex.ru/maps/43/kazan/?from=api-maps&ll=49.112172%2C55.791440&mode=routes&origin=jsapi_2_1_79&rtext=~55.791440%2C49.112172&rtt=auto&ruri=~ymapsbm1%3A%2F%2Fgeo%3Fdata%3DIgoN3XJEQhVvKl9C&z=16')(
-                        target='_blank'
-                    ) Построить маршрут
+    include components/map-contacts/map-contacts
